+++ gulpfile.js
@@ -39,15 +39,18 @@ const prettier = require('gulp-prettier');
 const postcss     = require('gulp-postcss');
 const reporter    = require('postcss-reporter');
 const syntax_scss = require('postcss-scss');
-const stylelint   = require('stylelint');
+const gulpStylelint = require('gulp-stylelint');
 
 // Image things
-
 const svgo = require('gulp-svgo');
 
 // JS Stuff
 const concat = require('gulp-concat');
 const uglify = require('gulp-uglify');
+// const sourcemaps = require('gulp-sourcemaps');
+const babel = require('gulp-babel');
+const rollup = require('gulp-better-rollup');
+const includePaths = require('rollup-plugin-includepaths');
 
 const patternPath = path.resolve(__dirname, 'components' );
 
@@ -64,8 +67,6 @@ const theo = require('theo')
 // -----------------------------------------------------------------------------
 
 gulp.task('css', function() {
-  const options = {
-  };
   const opts = {
     importer: [nodeModuleImport],
     includePaths: [
@@ -73,6 +74,7 @@ gulp.task('css', function() {
       path.resolve(__dirname, 'components/vf-sass-config/functions'),
       path.resolve(__dirname, 'components/vf-sass-config/mixins'),
       path.resolve(__dirname, 'components'),
+      path.resolve(__dirname, 'components/vf-form'),
       path.resolve(__dirname, 'components/vf-core-patterns'),
       path.resolve(__dirname, 'node_modules'),
     ]
@@ -84,9 +86,11 @@ gulp.task('css', function() {
     .on(
       'error',
       notify.onError(function(error) {
-        return 'Problem file : ' + error.message;
+        process.emit('exit') // this fails precommit, but allows guld dev to work
+        return 'Problem at file: ' + error.message;
       })
     )
+    .pipe(autoprefixer(autoprefixerOptions))
     .pipe(browserSync.stream())
     .pipe(sourcemaps.write())
     .pipe(rename(
@@ -105,92 +109,75 @@ gulp.task('css', function() {
 });
 
 // Sass Lint
-gulp.task("scss-lint", function() {
+gulp.task("scss-lint", function lintCssTask() {
 
-  // Stylelint config rules
-  var stylelintConfig = {
-    "plugins": [
-      "stylelint-order",
-      "stylelint-scss"
-    ],
-    "extends": [
-      "./node_modules/prettier-stylelint/config.js"
-    ],
-    "rules": {
-      "indentation": 2,
-      "string-quotes": "single",
-      "order/order": [
-        [
-        "custom-properties",
-        "dollar-variables",
-        {
-        type: "at-rule",
-        name: "include",
-        parameter: "set-*",
-        message: "Stop being lazy and flatten your classnames"
-        },
-        "declarations",
-        "rules",
-        {
-        	type: "at-rule",
-        	name: "media",
-        	hasBlock: true
-        }
-      ]
-      ],
-      "order/properties-alphabetical-order": true,
-      "block-no-empty": true,
-      "color-no-invalid-hex": true,
-      "declaration-colon-space-after": "always",
-      "declaration-colon-space-before": "never",
-      "function-comma-space-after": "always",
-      "media-feature-colon-space-after": "always",
-      "media-feature-colon-space-before": "never",
-      "media-feature-name-no-vendor-prefix": true,
-      "max-empty-lines": 3,
-      "max-nesting-depth": 3,
-      "selector-nested-pattern": [ "^(?!&__|&--|&-|&_).*", {
-        "message": "Stop being lazy and flatten your classnames",
-        "severity": "warning"
-      } ],
-      "number-leading-zero": "never",
-      "number-no-trailing-zeros": true,
-      "property-no-vendor-prefix": true,
-      "selector-list-comma-space-before": "never",
-      "selector-list-comma-newline-after": "always",
-      "string-quotes": "single",
-      "value-no-vendor-prefix": true,
-      // "indentation": [ 2, {
-      //   "except": ["block"],
-      //   "message": "Please use 2 spaces for indentation. Tabs make Stuart very grumpy.",
-      //   "severity": "warning"
-      // } ]
-    }
-  }
-
-  var processors = [
-    stylelint(stylelintConfig),
-    reporter({
-      clearMessages: true,
-      throwError: false,
-    })
-  ];
+  // For stylelint config rules see .stylelinrc
 
   return gulp
     .src(
-      ['components/**/vf-*.scss', '!components/**/index.scss', '!assets/**/*.scss']
+      ['components/**/embl-*.scss', 'components/**/vf-*.scss', '!components/**/index.scss', '!assets/**/*.scss']
     )
-    .pipe(postcss(processors, {syntax: syntax_scss}));
+    .pipe(gulpStylelint({
+      failAfterError: true,
+      reporters: [
+        {formatter: "string", console: true}
+      ]
+    }));
 });
 
 
 // -----------------------------------------------------------------------------
 // Scripts Tasks
 // -----------------------------------------------------------------------------
-gulp.task('scripts', function() {
-  return gulp
-    .src('./components/**/*.js')
-    .pipe(concat('scripts.js'))
+
+// Rollup all JS imports into CJS and babel them to ES5
+gulp.task('scripts:es5', function() {
+  let includePathOptions = {
+      include: {},
+      paths: [
+        path.resolve(__dirname, 'components'),
+        path.resolve(__dirname, 'components/vf-core-patterns'),
+        path.resolve(__dirname, 'components/vf-form'),
+      ],
+      external: ['vfTabs'],
+      extensions: ['.js']
+  };
+
+  return gulp.src('./components/vf-core/scripts.js')
+    // .pipe(sourcemaps.init())
+    .pipe(rollup({
+      // There is no `input` option as rollup integrates into the gulp pipeline
+      treeshake: false,
+      // external: ['vfTabs','vfBanner'],
+      plugins: [
+        babel({
+          "presets": [
+            [
+              "@babel/preset-env",
+              {
+                "targets": "> 0.25%, not dead, last 2 versions"
+              }
+            ]
+          ]
+        }),
+        includePaths(includePathOptions)
+      ]
+    }, {
+      // Rollups `sourcemap` option is unsupported. Use `gulp-sourcemaps` plugin instead
+      format: 'cjs',
+    }))
+    // inlining the sourcemap into the exported .js file
+    // .pipe(sourcemaps.write())
+    .pipe(gulp.dest('./public/scripts'));
+});
+
+
+// Eventually we'll want to support ES6 natively with ES5 as fallback, `scripts.es5.js`
+gulp.task('scripts:modern', function() {
+  return gulp.src('./components/vf-core/scripts.js')
+      .pipe(rename(function (path) {
+        path.extname = ".modern.js";
+      }))
     .pipe(gulp.dest('./public/scripts'));
 });
 
@@ -201,7 +188,6 @@ gulp.task('pattern-assets', function() {
   return gulp
     .src(['./components/**/**/assets/**/*'])
     .pipe(gulp.dest('./public/assets'));
-    // .pipe(gulp.dest('./assets'));
 });
 
 
@@ -318,6 +304,7 @@ var genCss = function (option) {
         path.resolve(__dirname, 'components/vf-sass-config/functions'),
         path.resolve(__dirname, 'components/vf-sass-config/mixins'),
         path.resolve(__dirname, 'components'),
+        path.resolve(__dirname, 'components/vf-form'),
         path.resolve(__dirname, 'components/vf-core-patterns'),
         path.resolve(__dirname, 'node_modules')
       ],
@@ -351,10 +338,13 @@ gulp.task('watch', function(done) {
   gulp.watch('./components/**/**/assets/*', gulp.series('images', 'pattern-assets')).on('change', reload);
 });
 
-
 // -----------------------------------------------------------------------------
 // Default Tasks
 // -----------------------------------------------------------------------------
+
+gulp.task('scripts', gulp.series(
+  'scripts:es5', 'scripts:modern'
+));
 
 // Build as a static site for CI
 gulp.task('build', gulp.series(
@@ -367,6 +357,10 @@ gulp.task('dev', gulp.parallel(
 
 gulp.task('tokens', gulp.parallel(
   'tokens:variables', 'tokens:typographic-scale', 'tokens:maps'
+));
+
+gulp.task('prepush-test', gulp.parallel(
+  'scss-lint', 'css'
 ));
 
 gulp.task('component', shell.task(