+++ components/vf-boilerplate-page/CHANGELOG.md
@@ -1,3 +1,24 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## 0.0.13 (2019-02-28)
+
+**Note:** Version bump only for package @visual-framework/vf-boilerplate-page
+
+
+
+
+
+## 0.0.12 (2019-02-28)
+
+**Note:** Version bump only for package @visual-framework/vf-boilerplate-page
+
+
+
+
+
 # 0.1.0
 
 ### Added