+++ components/vf-code-example/CHANGELOG.md
@@ -1,3 +1,32 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## 0.0.27 (2019-02-28)
+
+
+
+# 2.0.0-alpha.1 (2018-12-20)
+
+**Note:** Version bump only for package @visual-framework/vf-code-example
+
+
+
+
+
+## 0.0.26 (2019-02-28)
+
+
+
+# 2.0.0-alpha.1 (2018-12-20)
+
+**Note:** Version bump only for package @visual-framework/vf-code-example
+
+
+
+
+
 # 0.1.0
 
 ### Added